Enable latency measurement
The prior "live" deployment of WPT (w3c-test.org) was known to
occasionally exhibit unacceptable response latency.

Amazon Web Services offers measurement latency as part of the "Route 53
health check." Because Amazon charges an additional ($1 USD per month)
for each health check that enables this feature, it should be enabled
only on an opt-in basis.

Extend the "health check" module to allow consumers to enable latency
measurement, and configure the web-platform-tests.live project to enable
it.