Introduce additional health check
The web-platform-tests.live deployment temporarily becomes unavailable
via HTTPS. Add a health check to determine if the same can be said for
HTTP.

Update the Terraform module for AWS health checks to designate the
protocol in use. Because no other project in this repository reference
the module, this change does not affect any additional servers.