web-platform-tests.live: Honor non-existent domain
In commit 111587c777e4e05 we accommodated future changes to WPT's list
of available subdomains by introducing a "wildcard" DNS entry. This had
the unintended side-effect of enabling the domain which that project
explicitly endorses as non-existent
(nonexistent.web-platform-tests.live). Resolve this with a dedicated DNS
CNAME record, ensuring that requests to that particular domain do not
resolve.