Decommission pr-dashboard project (#11)
The configuration which integrated this project with the
web-platform-tests was disabled on 2018-05-18 [1]. The application's
content is now only useful for historical reference, and no dynamism or
computation is necessary to fulfill that purpose. A snapshot of the
application's current content has been created and made available in the
project's source code repository. Bocoup will maintain that as a static
collection of HTML files published to the web.

Remove the infrastructure originally provisioned to power the dynamic
web application.

[1] https://github.com/web-platform-tests/wpt/issues/10923